# Speedup parallel threads by removing superfluos synchronization