--- a/news/123.bugfix
+++ b/news/123.bugfix
@@ -1,0 +1,1 @@
+Speedup (~1.6x) parallel thread execution by removing superfluos synchronization [jensens]