While debugging the latest problems of @tisto s Volto project I found possible (meanwhile) superfluos thread synchonizations. Volto in some scenarios creates lots of parallel small and short requests via `plone.restapi`. Blocking makes the thread idle and waiting where not needed, even if this is only a very short time (in low ms area) this happens often an accumulates. This is relevant for high traffic classic Plone sites too.

In Py-Spy with a real world Volto project the main-page timings before are:

```
  0.00%   0.00%   0.000s    21.08s   render (plone/restapi/services/__init__.py:21)
  0.00%   0.00%   0.000s    19.53s   reply (plone/restapi/services/content/get.py:18)
  0.00%   0.00%   0.000s    18.79s   __call__ (plone/restapi/serializer/dxcontent.py:125)
  0.00%   0.00%   0.000s    14.30s   __call__ (plone/restapi/serializer/dxcontent.py:83)
```
after this changes:
```
  0.00%   0.00%   0.000s    13.59s   render (plone/restapi/services/__init__.py:21)
  0.00%   0.00%   0.010s    12.41s   reply (plone/restapi/services/content/get.py:18)
  0.00%   0.00%   0.000s    11.72s   __call__ (plone/restapi/serializer/dxcontent.py:125)
  0.00%   0.00%   0.000s     8.87s   __call__ (plone/restapi/serializer/dxcontent.py:83)
```