@datakurre 2019-12-27
@jensens Why superfluos?

I’ve also seen these blocking and I’m excited that these can be removed, but why is it safe?