@jensens 2019-12-27
@datakurre The caches are all stored on the `_v_..` attributes of the FTI and on the request object, so both are thread safe by nature. The only non-threadsafe is the invalidation counter.