@datakurre 2019-12-27
@jensens Thanks for explaining. Another great catch then!