@jensens 2019-12-27
Do not merge yet, I just found that the invalidation of `_v_` as in this implementation wont work perfectly.
A `_v_` lives in its ZEO-cache. Invalidations would only invalidate the current cache. Thus we need a common variable invalidating the other caches. I have an idea how to do so, stay tuned.