@jensens 2019-12-27
No, all fine. I forgot to commit the transaction. Then the other ZEO has the `_v_` attribute too. ZODB documentation is here a bit thin.