@davisagli 2019-12-28
@jensens nice find! I think these locks were left over from the original cache implementation which did not use persistent storage: https://github.com/plone/plone.dexterity/commit/f107f0d2ed82a5adf126b5fa12e4e409a88aa772